+++ backend/internal/v1/v1_users/users.go
@@ -1,1 +1,1 @@
-package v1users
+package v1_users
